etmem: psi fb: add wildcard resolution for cg_path field
cg_path indicates the path of the Cgroup to be swapped out.

In earlier versions, only one cg_path can be added to a task.
If multiple containers are added to a cgroup path in
a scenario, it is difficult to add cg_path.

To solve the preceding problem, the wildcard parsing
function for cg_path is added.

Example: cg_path=isulad/*
In this case, all level-2 directories in the
/sys/fs/cgroup/cpuacct/isulad/ directory are added
to the task for scanning and swapping out.

Signed-off-by: liubo <liubo254@huawei.com>