cmake: Pacify CMake with no functional changes
While running cmake on the lastest openEuler 24.03, I ran into,

---------------------------------------------------------------------
[root@loaclhost build]# cmake ..
CMake Warning (dev) at CMakeLists.txt:16 (project):
  cmake_minimum_required() should be called prior to this top-level project()
  call.  Please see the cmake-commands(7) manual for usage documentation of
  both commands.
This warning is for project developers.  Use -Wno-dev to suppress it.
---------------------------------------------------------------------

and I noticed that the documentation [1] online pointed out this in detail
as follows,

    Note: Call the cmake_minimum_required() command at the beginning of the
          top-level CMakeLists.txt file even before calling the project()
          command.

Therefore, fix this by simply putting cmake_minimum_required() at the top of
the CMakeList file.

Tested on x86_64 openEuler 20.03, 22.04, and 24.03.

[1] https://cmake.org/cmake/help/latest/command/cmake_minimum_required.html

Signed-off-by: Enze Li <lienze@kylinos.cn>