etmem: Mask the '-s' option of the linker when enabling CONFIG_DEBUG
A couple days ago, when I tried to debug etmem with GDB, it said to me
that no debugging symbols found in etmem, even though I used the
-DCONFIG_DEBUG=y option.

Upon digging further, I realized that the root cause is that there's a
linker options '-s' in set_target_properties(), which means the linker
will omit all symbol information from the output file.

Thus, even with the addition of the '-g' option, the generated symbols
will eventually be cleared.

This patch fixes this issue by masking '-s' option of the linker, as
well as the same issue in memRouter.

Tested on x86_64 openEuler 20.03, 22.03 and 24.03.

Signed-off-by: Enze Li <lienze@kylinos.cn>