etmem: Fix build issue caused by undefined pmu_params structure
I ran into the following error when building etmem with the default
configuration.

---------------------------------------------------------------------
[  1%] Building C object etmem/CMakeFiles/etmemd.dir/src/etmemd_src/etmemd.c.o
In file included from /root/dev/etmem/etmem/src/etmemd_src/etmemd.c:23:
/root/dev/etmem/etmem/inc/etmemd_inc/etmemd_scan.h:108:50: error: ‘struct pmu_params’ declared inside parame
ter list will not be visible outside of this definition or declaration [-Werror]
  108 | inline struct page_refs *etmemd_do_sample(struct pmu_params *pmu_params)
      |                                                  ^~~~~~~~~~
/root/dev/etmem/etmem/inc/etmemd_inc/etmemd_scan.h:112:39: error: ‘struct pmu_params’ declared inside parame
ter list will not be visible outside of this definition or declaration [-Werror]
  112 | inline void etmemd_stop_sample(struct pmu_params *pmu_params)
      |                                       ^~~~~~~~~~
---------------------------------------------------------------------

The root cause of this issue is that in this commit,

  commit 012fec939e77e9c362f3b17216d2efb7d9d67647
  Date:   Sat Sep 21 22:15:32 2024 +0800

    Adapt PMU sampling function to cslide

the pmu_params structure was moved to the etmemd_scan.h which is a common
header file, but from that point, this structure will only be used by PMU. In
practice, however, this structure will be used by other functions as well.

Fixing this is as simple as placing this structure outside of the ENABLE_PMU
macro.

Tested by rebuilding on x86_64 openEuler 20.03, 22.03 and 24.03.

Fixes: 012fec939e77 ("Adapt PMU sampling function to cslide")
Signed-off-by: Enze Li <lienze@kylinos.cn>